backend/src/requirements.txt
asgiref==3.2.10
Django==3.1.13
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
google-auth==1.30.1
django-cors-headers==3.7.0
